backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.0
pydantic>=2.10.0
python-multipart>=0.0.7
aiofiles>=24.0.0
requests>=2.32.0
httpx>=0.26.0
pgvector>=0.3.0
psycopg2-binary>=2.9.0
pytest>=7.4.4
pytest-asyncio>=0.23.0
